requirements.txt
flask==2.0.1
flask-socketio==5.1.1
python-dotenv==0.19.0
openai==1.3.0
gunicorn==21.2.0
eventlet==0.33.3
rich==13.7.0
colorama==0.4.6
python-socketio==5.11.1 